another solution: Implement Queue Using Stack
